+++ python/packages/durabletask/agent_framework_durabletask/_workflows/serialization.py
@@ -1,19 +1,32 @@
 # Copyright (c) Microsoft. All rights reserved.
 
-"""Serialization utilities for workflow execution.
+"""Internal serialization helpers for workflow execution.
 
-This module provides thin wrappers around the core checkpoint encoding system
-(encode_checkpoint_value / decode_checkpoint_value) from agent_framework._workflows.
+These helpers are framework-internal plumbing for moving typed objects across
+durable orchestration/activity boundaries. They are **not** part of the public
+API and must not be called by application code.
 
-The core checkpoint encoding uses pickle + base64 for type-safe roundtripping of
-arbitrary Python objects (dataclasses, Pydantic models, Message, etc.) while
-keeping JSON-native types (str, int, float, bool, None) as-is.
+They wrap the core checkpoint codec (``encode_checkpoint_value`` /
+``decode_checkpoint_value`` from ``agent_framework._workflows``), which uses
+pickle + base64 to round-trip arbitrary Python objects (dataclasses, Pydantic
+models, ``Message``, ...) while leaving JSON-native types (str, int, float,
+bool, None) as-is.
 
-This module adds:
-- serialize_value / deserialize_value: convenience aliases for encode/decode
-- reconstruct_to_type: for HITL responses where external data (without type markers)
-  needs to be reconstructed to a known type
-- resolve_type: resolves 'module:class' type keys to Python types
+Because that codec can unpickle objects, every value that crosses an external
+trust boundary -- HTTP request bodies and HITL responses raised as external
+events -- is sanitized by the framework with :func:`strip_pickle_markers`
+*before* it can reach these helpers. Application code never has to perform that
+sanitization itself: the orchestrator, the activity body, and the HTTP entry
+points already do it at the boundary. See
+:mod:`agent_framework._workflows._checkpoint_encoding` for the full security model.
+
+Contents:
+- ``_serialize_value`` / ``_deserialize_value``: internal codec aliases for encode/decode.
+- ``reconstruct_to_type``: rebuilds HITL response data (which arrives without type
+  markers) to a known type.
+- ``resolve_type``: resolves 'module:class' type keys to Python types.
+- ``strip_pickle_markers``: the framework's trust-boundary defense that neutralizes
+  attacker-injected pickle/type markers.
 """
 
 from __future__ import annotations
@@ -75,9 +88,10 @@ def strip_pickle_markers(data: Any) -> Any:
     that contains either marker key with ``None``, neutralizing the attack
     vector while leaving all other data untouched.
 
-    It **must** be called on every value that originates from an untrusted
-    source (e.g. ``req.get_json()``) *before* the value is passed to
-    ``deserialize_value`` / ``decode_checkpoint_value``.
+    The framework applies this at every external trust boundary -- HTTP request
+    bodies and HITL responses raised as external events -- before the value can
+    reach the internal codec (:func:`_deserialize_value` /
+    ``decode_checkpoint_value``). Application code does not need to call it.
     """
     if isinstance(data, dict):
         if _PICKLE_MARKER in data or _TYPE_MARKER in data:
@@ -98,11 +112,12 @@ def strip_pickle_markers(data: Any) -> Any:
 # ============================================================================
 
 
-def serialize_value(value: Any) -> Any:
-    """Serialize a value for JSON-compatible cross-activity communication.
+def _serialize_value(value: Any) -> Any:
+    """Encode a value for JSON-compatible cross-activity communication (internal).
 
-    Delegates to core checkpoint encoding which uses pickle + base64 for
-    non-JSON-native types (dataclasses, Pydantic models, Message, etc.).
+    Framework-internal codec. Delegates to core checkpoint encoding which uses
+    pickle + base64 for non-JSON-native types (dataclasses, Pydantic models,
+    Message, etc.). Not part of the public API.
 
     Args:
         value: Any Python value (primitive, dataclass, Pydantic model, Message, etc.)
@@ -113,11 +128,14 @@ def serialize_value(value: Any) -> Any:
     return encode_checkpoint_value(value)
 
 
-def deserialize_value(value: Any) -> Any:
-    """Deserialize a value previously serialized with serialize_value().
+def _deserialize_value(value: Any) -> Any:
+    """Decode a value previously encoded with :func:`_serialize_value` (internal).
 
-    Delegates to core checkpoint decoding which unpickles base64-encoded values
-    and verifies type integrity.
+    Framework-internal codec. Delegates to core checkpoint decoding which
+    unpickles base64-encoded values and verifies type integrity. Not part of the
+    public API: callers only ever hand it values that the framework produced
+    itself or that have already passed the :func:`strip_pickle_markers` trust
+    boundary, so untrusted markers can never reach ``pickle.loads()`` here.
 
     Args:
         value: The serialized data (dict with pickle markers, list, or primitive)
@@ -131,7 +149,7 @@ def deserialize_value(value: Any) -> Any:
 def deserialize_workflow_output(output: Any) -> Any:
     """Reconstruct the workflow outputs produced by the shared activity.
 
-    Each value an executor yields is encoded with :func:`serialize_value` before
+    Each value an executor yields is encoded with :func:`_serialize_value` before
     it reaches the orchestrator, so typed objects (dataclasses, Pydantic models,
     ``AgentResponse``, ...) are stored as checkpoint-marker dicts. This reverses
     that encoding so callers receive the original objects.
@@ -152,7 +170,7 @@ def deserialize_workflow_output(output: Any) -> Any:
         The output with every checkpoint-encoded value reconstructed; primitives
         and plain JSON structures pass through unchanged.
     """
-    return deserialize_value(output)
+    return _deserialize_value(output)
 
 
 # ============================================================================
@@ -172,7 +190,7 @@ def serialize_workflow_event(event: WorkflowEvent[Any]) -> dict[str, Any]:
 
     Carries a workflow event from the durable activity, through the orchestration
     custom status, to a streaming client. The data payload is encoded with
-    :func:`serialize_value` so typed objects survive the round trip;
+    :func:`_serialize_value` so typed objects survive the round trip;
     :func:`deserialize_workflow_event` reverses it into a ``WorkflowEvent`` so
     callers never handle checkpoint-marker dicts directly.
 
@@ -187,7 +205,7 @@ def serialize_workflow_event(event: WorkflowEvent[Any]) -> dict[str, Any]:
     if event.executor_id is not None:
         serialized["executor_id"] = event.executor_id
     if event.data is not None:
-        serialized["data"] = serialize_value(event.data)
+        serialized["data"] = _serialize_value(event.data)
     if event.type == "request_info":
         # request_type is omitted: deserialize_workflow_event rebuilds the event via
         # WorkflowEvent.request_info, which derives it from the data payload.
@@ -202,7 +220,7 @@ def deserialize_workflow_event(serialized: dict[str, Any]) -> WorkflowEvent[Any]
 
     ``serialized`` must originate from the workflow's own orchestration custom
     status (trusted durable storage); its encoded payload is decoded with
-    :func:`deserialize_value`. Never pass untrusted external input here.
+    :func:`_deserialize_value`. Never pass untrusted external input here.
 
     Args:
         serialized: A dict previously produced by :func:`serialize_workflow_event`,
@@ -212,7 +230,7 @@ def deserialize_workflow_event(serialized: dict[str, Any]) -> WorkflowEvent[Any]
         The reconstructed workflow event with its data payload restored.
     """
     event_type = cast(WorkflowEventType, serialized["type"])
-    payload = deserialize_value(serialized["data"]) if "data" in serialized else None
+    payload = _deserialize_value(serialized["data"]) if "data" in serialized else None
 
     if event_type == "request_info":
         response_key = serialized.get("response_type")
@@ -245,7 +263,7 @@ def reconstruct_to_type(value: Any, target_type: type) -> Any:
 
     Tries strategies in order:
     1. Return as-is if already the correct type
-    2. deserialize_value (for data with any type markers)
+    2. _deserialize_value (for data with any type markers)
     3. Pydantic model_validate (for Pydantic models)
     4. Dataclass constructor (for dataclasses)
 
@@ -270,7 +288,7 @@ def reconstruct_to_type(value: Any, target_type: type) -> Any:
     # NOTE: This function is general-purpose.  Callers that handle untrusted
     # data (e.g. HITL responses) MUST call strip_pickle_markers() before
     # passing data here.  See _deserialize_hitl_response in orchestrator.py.
-    decoded = deserialize_value(value)
+    decoded = _deserialize_value(value)
     if not isinstance(decoded, dict):
         return decoded
